add structs.go, fix auth test
