+++ cmd/client/cmd/structs.go
@@ -1,0 +1,16 @@
+package cmd
+
+type newUserInfo struct {
+	Username string `json:"username"`
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
+type tokenInfo struct {
+	Token string `json:"token"`
+}
+
+type loginInfo struct {
+	Login    string `json:"login"`
+	Password string `json:"password"`
+}
